Properly refactor to PhysicsTime
