+++ core/src/lib.rs
@@ -10,13 +10,13 @@ use bevy::prelude::*;
 pub use constraints::RotationConstraints;
 pub use ext::*;
 pub use gravity::Gravity;
-pub use physics_controller::PhysicsTime;
+pub use physics_time::PhysicsTime;
 pub use velocity::{Acceleration, AxisAngle, Velocity};
 
 mod constraints;
 pub mod ext;
 mod gravity;
-mod physics_controller;
+mod physics_time;
 pub mod utils;
 mod velocity;
